move common_koji to module_utils
This gets rid of the PYTHONPATH hack.
