+++ README.rst
@@ -184,23 +184,26 @@ no explicit ``koji:`` argument::
 
    KOJI_PROFILE=kojidev ansible-playbook -v my-koji-playbook.yaml
 
-Python paths
-------------
+File paths
+----------
 
-These modules import from other files in the ``library`` directory. If you get
-``ImportError`` when using these modules,  set the ``PYTHONPATH`` environment
-variable to this ``library`` directory.
+These modules import ``common_koji`` from the ``module_utils`` directory.
+
+One easy way to arrange your Ansible files is to symlink the ``library`` and
+``module_utils`` directories into the directory with your playbook.
 
 For example, if you have a ``koji.yml`` playbook that you run with
-``ansible-playbook``, it should live alongside this ``library`` directory::
+``ansible-playbook``, it should live alongside these ``library`` and
+``module_utils`` directories::
 
     top
     ├── koji.yml
+    ├── module_utils
     └── library
 
 and you should run the playbook like so::
 
-   PYTHONPATH=library ansible-playbook koji.yml
+   ansible-playbook koji.yml
 
 
 TODO
